split hash functionality into separate file
