+++ pkg/cluster/deploy.go
@@ -2,14 +2,8 @@ package cluster
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/base64"
-	"encoding/json"
-
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2018-06-01/compute"
 
 	"github.com/openshift/openshift-azure/pkg/api"
-	"github.com/openshift/openshift-azure/pkg/arm"
 	"github.com/openshift/openshift-azure/pkg/config"
 	"github.com/openshift/openshift-azure/pkg/util/azureclient/storage"
 	"github.com/openshift/openshift-azure/pkg/util/managedcluster"
@@ -72,46 +66,6 @@ func (u *simpleUpgrader) Deploy(ctx context.Context, cs *api.OpenShiftManagedClu
 
 type scalesetName string
 type instanceName string
-type hash string
-
-func hashVMSS(vmss *compute.VirtualMachineScaleSet) (hash, error) {
-	// cleanup capacity so that no unnecessary VM rotations are going to occur
-	// because of a scale up/down.
-	if vmss.Sku != nil {
-		vmss.Sku.Capacity = nil
-	}
-
-	data, err := json.Marshal(vmss)
-	if err != nil {
-		return "", err
-	}
-
-	hf := sha256.New()
-	hf.Write(data)
-
-	return hash(base64.StdEncoding.EncodeToString(hf.Sum(nil))), nil
-}
-
-// hashScaleSets returns the set of desired state scale set hashes
-func (u *simpleUpgrader) hashScaleSets(cs *api.OpenShiftManagedCluster) (map[scalesetName]hash, error) {
-	ssHashes := map[scalesetName]hash{}
-
-	for _, app := range cs.Properties.AgentPoolProfiles {
-		vmss, err := arm.Vmss(&u.pluginConfig, cs, &app, "") // TODO: backupBlob is rather a layering violation here
-		if err != nil {
-			return nil, err
-		}
-
-		h, err := hashVMSS(vmss)
-		if err != nil {
-			return nil, err
-		}
-
-		ssHashes[scalesetName(*vmss.Name)] = h
-	}
-
-	return ssHashes, nil
-}
 
 func (u *simpleUpgrader) initializeUpdateBlob(cs *api.OpenShiftManagedCluster, ssHashes map[scalesetName]hash) error {
 	blob := newUpdateBlob()
